+++ EssentialFeed/EssentialFeediOSTests/FeedSnapshotTests.swift
@@ -78,55 +78,6 @@ final class FeedSnapshotTests: XCTestCase {
         ]
     }
     
-    private func record(snapshot: UIImage, name: String, file: StaticString = #filePath, line: UInt = #line) {
-        
-        let snapshotData = makeSnapshotData(for: snapshot, file: file, line: line)
-        let snapshotURL = makeSnapshotURL(named: name, file: file)
-        
-        do {
-            try FileManager.default.createDirectory(
-                at: snapshotURL.deletingLastPathComponent(),
-                withIntermediateDirectories: true)
-            
-            try snapshotData?.write(to: snapshotURL)
-        } catch {
-            XCTFail("Failed to record snapshot with error: \(error)", file: file, line: line)
-        }
-    }
-    
-    private func assert(snapshot: UIImage, name: String, file: StaticString = #filePath, line: UInt = #line) {
-        
-        let snapshotData = makeSnapshotData(for: snapshot, file: file, line: line)
-        let snapshotURL = makeSnapshotURL(named: name, file: file)
-        
-        guard let storedSnapshotData = try? Data(contentsOf: snapshotURL) else {
-            return XCTFail("Failed to load stored snapshot at URL: \(snapshotURL). Use the `record` method to store a snapshot before asserting.", file: file, line: line)
-        }
-        if storedSnapshotData != snapshotData {
-            let temporarySnapshotURL = URL(fileURLWithPath: NSTemporaryDirectory(), isDirectory: true)
-                .appendingPathComponent(snapshotURL.lastPathComponent)
-            
-            try? snapshotData?.write(to: temporarySnapshotURL)
-            
-            XCTFail("New snapshot does not match stored snapshot. New snapshot URL: \(temporarySnapshotURL), Stored snapshot URL: \(snapshotURL)", file: file, line: line)
-        }
-    }
-    
-    private func makeSnapshotData(for snapshot: UIImage, file: StaticString = #filePath, line: UInt = #line) -> Data? {
-        guard let snapshotData = snapshot.pngData() else {
-            XCTFail("Failed to generate PNG data representation from snapshot", file: file, line: line)
-            return nil
-        }
-        return snapshotData
-    }
-    
-    private func makeSnapshotURL(named name: String, file: StaticString = #filePath) -> URL {
-        URL(fileURLWithPath: String(describing: file))
-            .deletingLastPathComponent()
-            .appendingPathComponent("snapshots")
-            .appendingPathComponent("\(name).png")
-    }
-    
 }
 
 private extension FeedViewController {
